Inproved API design
