+++ examples/pipeline/main.log
@@ -1,13 +1,15 @@
    Compiling spirq v0.1.0 (C:\Users\PENGUINLIONG\Repositories\spirq)
-    Finished dev [unoptimized + debuginfo] target(s) in 1.38s
+    Finished dev [unoptimized + debuginfo] target(s) in 1.74s
      Running `target\debug\examples\pipeline.exe`
-[2019-12-31T13:43:54Z INFO  pipeline] collected spirvs: ["uniform-pbr.vert", "uniform-pbr.frag"]
-[2019-12-31T13:43:54Z INFO  pipeline] push_constant[model_view]: offset=Some(0), ty=mat4x4<f32>
-[2019-12-31T13:43:54Z INFO  pipeline] push_constant[view_proj]: offset=Some(64), ty=mat4x4<f32>
-[2019-12-31T13:43:54Z INFO  pipeline] mat.fdsa.1: offset=Some(80), ty=mat3x3<f32>
-[2019-12-31T13:43:54Z INFO  pipeline] someImage: offset=None, ty=subpassData
-[2019-12-31T13:43:54Z INFO  pipeline] imgggg: offset=None, ty=sampler2D
-[2019-12-31T13:43:54Z INFO  pipeline] -- buffer sizing:
-[2019-12-31T13:43:54Z INFO  pipeline] (set=0, bind=1): nbyte=Some(128)
-[2019-12-31T13:43:54Z INFO  pipeline] (push_constant): nbyte=Some(16)
-[2019-12-31T13:43:54Z INFO  pipeline] (set=1, bind=0): nbyte=None
+[2020-01-01T16:48:05Z INFO  pipeline] collected spirvs: ["uniform-pbr.frag", "uniform-pbr.vert"]
+[2020-01-01T16:48:05Z INFO  pipeline] push_constant[model_view]: DescriptorResolution { desc_bind: (push_constant), desc_ty: { randomColor: vec4<f32>, model_view: mat4x4<f32>, view_proj: mat4x4<f32> }, member_var_res: Some(MemberVariableResolution { offset: 0, ty: mat4x4<f32> }) }
+[2020-01-01T16:48:05Z INFO  pipeline] push_constant[view_proj]: DescriptorResolution { desc_bind: (push_constant), desc_ty: { randomColor: vec4<f32>, model_view: mat4x4<f32>, view_proj: mat4x4<f32> }, member_var_res: Some(MemberVariableResolution { offset: 64, ty: mat4x4<f32> }) }
+[2020-01-01T16:48:05Z INFO  pipeline] mat.fdsa.1: DescriptorResolution { desc_bind: (set=0, bind=1), desc_ty: 2x{ albedo: vec3<f32>, metalicity: f32, roughness: f32, fdsa: [mat3x3<f32>; 2] }, member_var_res: Some(MemberVariableResolution { offset: 80, ty: mat3x3<f32> }) }
+[2020-01-01T16:48:05Z INFO  pipeline] someImage: DescriptorResolution { desc_bind: (set=0, bind=2), desc_ty: subpassData[0], member_var_res: None }
+[2020-01-01T16:48:05Z INFO  pipeline] imgggg: DescriptorResolution { desc_bind: (set=0, bind=3), desc_ty: sampler2D, member_var_res: None }
+[2020-01-01T16:48:05Z INFO  pipeline] -- buffer sizing:
+[2020-01-01T16:48:05Z INFO  pipeline] (set=1, bind=0): nbyte=Some(16)
+[2020-01-01T16:48:05Z INFO  pipeline] (set=0, bind=1): nbyte=Some(128)
+[2020-01-01T16:48:05Z INFO  pipeline] (push_constant): nbyte=Some(128)
+[2020-01-01T16:48:05Z INFO  pipeline] (set=0, bind=3): nbyte=None
+[2020-01-01T16:48:05Z INFO  pipeline] (set=0, bind=2): nbyte=None
