+++ examples/query/main.log
@@ -1,13 +1,13 @@
-Compiling spirq v0.1.0 (C:\Users\PENGUINLIONG\Repositories\spirq)
-    Finished dev [unoptimized + debuginfo] target(s) in 1.47s
+   Compiling spirq v0.1.0 (C:\Users\PENGUINLIONG\Repositories\spirq)
+    Finished dev [unoptimized + debuginfo] target(s) in 1.06s
      Running `target\debug\examples\query.exe`
-[2019-12-28T09:56:55Z INFO  query] collected spirvs: ["uniform-pbr.frag", "uniform-pbr.vert"]
-[2019-12-28T09:56:55Z INFO  query] [
+[2020-01-01T16:50:43Z INFO  query] collected spirvs: ["uniform-pbr.frag", "uniform-pbr.vert"]
+[2020-01-01T16:50:43Z INFO  query] [
         main {
             inputs: {
-                0: vec4<f32>,
                 1: vec4<f32>,
                 2: vec2<f32>,
+                0: vec4<f32>,
             },
             outputs: {},
             descriptors: {
@@ -15,23 +15,23 @@ Compiling spirq v0.1.0 (C:\Users\PENGUINLIONG\Repositories\spirq)
             },
         },
     ]
-[2019-12-28T09:56:55Z INFO  query] push_constant[model_view]: offset=Some(0), ty=mat4x4<f32>
-[2019-12-28T09:56:55Z INFO  query] push_constant[view_proj]: offset=Some(64), ty=mat4x4<f32>
-[2019-12-28T09:56:55Z INFO  query] [
+[2020-01-01T16:50:43Z INFO  query] push_constant[model_view]: DescriptorResolution { desc_bind: (push_constant), desc_ty: { model_view: mat4x4<f32>, view_proj: mat4x4<f32> }, member_var_res: Some(MemberVariableResolution { offset: 0, ty: mat4x4<f32> }) }
+[2020-01-01T16:50:43Z INFO  query] push_constant[view_proj]: DescriptorResolution { desc_bind: (push_constant), desc_ty: { model_view: mat4x4<f32>, view_proj: mat4x4<f32> }, member_var_res: Some(MemberVariableResolution { offset: 64, ty: mat4x4<f32> }) }
+[2020-01-01T16:50:43Z INFO  query] [
         main {
             inputs: {},
             outputs: {
                 0: vec4<f32>,
             },
             descriptors: {
-                (set=1, bind=0): [buffer: { cam_pos: vec4<f32>, lights: [{ pos: vec3<f32>, color: vec3<f32> } }; 2],
-                (push_constant): { randomColor: vec4<f32> },
+                (set=0, bind=1): 2x{ albedo: vec3<f32>, metalicity: f32, roughness: f32, fdsa: [mat3x3<f32>; 2] },
                 (set=0, bind=2): subpassData[0],
-                (set=0, bind=1): [uniform: { albedo: vec3<f32>, metalicity: f32, roughness: f32, fdsa: [mat3x3<f32>; 2] }; 2],
+                (push_constant): { randomColor: vec4<f32> },
                 (set=0, bind=3): sampler2D,
+                (set=1, bind=0): 2x{ cam_pos: vec4<f32>, lights: [{ pos: vec3<f32>, color: vec3<f32> } },
             },
         },
     ]
-[2019-12-28T09:56:55Z INFO  query] mat.fdsa.1: offset=Some(80), ty=mat3x3<f32>
-[2019-12-28T09:56:55Z INFO  query] someImage: offset=None, ty=subpassData
-[2019-12-28T09:56:55Z INFO  query] imgggg: offset=None, ty=sampler2D
+[2020-01-01T16:50:43Z INFO  query] mat.fdsa.1: DescriptorResolution { desc_bind: (set=0, bind=1), desc_ty: 2x{ albedo: vec3<f32>, metalicity: f32, roughness: f32, fdsa: [mat3x3<f32>; 2] }, member_var_res: Some(MemberVariableResolution { offset: 80, ty: mat3x3<f32> }) }
+[2020-01-01T16:50:43Z INFO  query] someImage: DescriptorResolution { desc_bind: (set=0, bind=2), desc_ty: subpassData[0], member_var_res: None }
+[2020-01-01T16:50:43Z INFO  query] imgggg: DescriptorResolution { desc_bind: (set=0, bind=3), desc_ty: sampler2D, member_var_res: None }
